Refactor model classes to enhance type handling and improve lineage representation with new DataTable, DataColumn, and ColumnLineage models.
